+++ resource/IPCA/src/ipcavariant.cpp
@@ -429,7 +429,7 @@ IPCAStatus IPCA_CALL IPCAPropertyBagGetValuePropertyBagArray(
     // rollback if any failure.
     if (i != ocrepCount)
     {
-        for (int x = 0; x < i; x++)
+        for (size_t x = 0; x < i; x++)
         {
             IPCAPropertyBagDestroy(*value[x]);
         }
